+++ test/compojure/api/meta_test.clj
@@ -1157,138 +1157,142 @@
       (dorun (repeatedly 10 exercise))
       (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "dynamic context where schema is bound outside context"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (let [s s/Any]
                   (context
                     "" []
                     :dynamic true
                     (GET "/ping" []
                          ;;TODO could lift this since the locals occur outside the context
-                         :body [body (do (swap! times inc) s)]
+                         :body-params [field :- (record :field s/Str)
+                                       field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                       & foo :- {(record :extra-keys s/Keyword)
+                                                 (record :extra-vals s/Keyword)} :as all]
                          (ok "kikka"))))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:body-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 11 @times))))
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "dynamic context that binds req and uses it in schema"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" req
                   (GET "/ping" req
-                       :body [body (do (swap! times inc)
-                                       ;; should never lift this since it refers to request
-                                       (second [req s/Any]))]
+                       :body-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:body-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 11 @times))))
-  (testing "bind :body in static context"
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
+  (testing "bind :body-params in static context"
     (is-thrown-with-msg?
       AssertionError
       #"cannot be :static"
       (eval `(context
                "" []
                :static true
-               :body [body (do (swap! times update :outer inc)
-                               s/Any)]
+               :body-params [field :- (record :field s/Str)
+                             field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                             & foo :- {(record :extra-keys s/Keyword)
+                                       (record :extra-vals s/Keyword)} :as all]
                (GET "/ping" req
                     :body [body (do (swap! times update :inner inc)
                                     s/Any)]
                     (ok "kikka"))))))
-  (testing "bind :body in dynamic context"
-    (let [times (atom {:outer 0 :inner 0})
+  (testing "bind :body-params in dynamic context"
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" []
                   :dynamic true
-                  :body [body (do (swap! times update :outer inc)
-                                  s/Any)]
                   (GET "/ping" req
-                       :body [body (do (swap! times update :inner inc)
-                                       s/Any)]
+                       :body-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:body-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= {:outer 1 :inner 1} @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= {:outer 1 :inner 11} @times))))
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "idea for lifting impl"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (let [rs (GET "/ping" req
-                              :body [body (do (swap! times inc)
-                                              s/Any)]
+                              :body-params [field :- (record :field s/Str)
+                                            field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                            & foo :- {(record :extra-keys s/Keyword)
+                                                      (record :extra-vals s/Keyword)} :as all]
                               (ok "kikka"))]
                   (context
                     "" []
                     :dynamic true
                     rs))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:body-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 1 @times)))))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 11} @times)))))
 
 (deftest form-params-double-eval-test
-  (testing "no :body-params double expansion"
+  (testing "no :form-params double expansion"
     (is-expands (GET "/ping" []
                      :form-params [field :- EXPENSIVE, field2, {default :- s/Int (inc 42)} & foo :- {s/Keyword s/Keyword} :as all]
                      (ok "kikka"))
-                '(compojure.api.routes/map->Route
-                   {:path "/ping",
-                    :method :get,
-                    :info
-                    (compojure.api.meta/merge-parameters
-                      {:public
-                       {:parameters
-                        {:formData
-                         {s/Keyword s/Keyword,
-                          :field EXPENSIVE,
-                          :field2 schema.core/Any,
-                          (clojure.core/with-meta
-                            (schema.core/optional-key :default)
-                            {:default '(inc 42)})
-                          s/Int}},
-                        :consumes ["application/x-www-form-urlencoded"]}}),
-                    :handler
-                    (compojure.core/make-route
-                      :get
-                      {:__record__ "clout.core.CompiledRoute",
-                       :source "/ping",
-                       :re (clojure.core/re-pattern "/ping"),
-                       :keys [],
-                       :absolute? false}
-                      (clojure.core/fn
-                        [?request]
-                        (compojure.core/let-request
-                          [[:as +compojure-api-request+] ?request]
-                          (plumbing.core/letk
-                            [[field
-                              :-
-                              EXPENSIVE
-                              field2
-                              {default :-, s/Int (inc 42)}
-                              :&
-                              foo
-                              :-
-                              #:s{Keyword s/Keyword}
-                              :as
-                              all]
-                             (compojure.api.coercion/coerce-request!
-                               {s/Keyword s/Keyword,
-                                :field EXPENSIVE,
-                                :field2 schema.core/Any,
-                                (clojure.core/with-meta
-                                  (schema.core/optional-key :default)
-                                  {:default '(inc 42)})
-                                s/Int}
-                               :form-params
-                               :string
-                               true
-                               false
-                               +compojure-api-request+)]
-                            (do (ok "kikka"))))))})))
+                '(clojure.core/let
+                   [?form-params-schema108882
+                    {s/Keyword s/Keyword,
+                     :field EXPENSIVE,
+                     :field2 schema.core/Any,
+                     (clojure.core/with-meta
+                       (schema.core/optional-key :default)
+                       {:default '(inc 42)})
+                     s/Int}]
+                   (compojure.api.routes/map->Route
+                     {:path "/ping",
+                      :method :get,
+                      :info
+                      (compojure.api.meta/merge-parameters
+                        {:public
+                         {:parameters {:formData ?form-params-schema108882},
+                          :consumes ["application/x-www-form-urlencoded"]}}),
+                      :handler
+                      (compojure.core/make-route
+                        :get
+                        {:__record__ "clout.core.CompiledRoute",
+                         :source "/ping",
+                         :re (clojure.core/re-pattern "/ping"),
+                         :keys [],
+                         :absolute? false}
+                        (clojure.core/fn
+                          [?request__3574__auto__]
+                          (compojure.core/let-request
+                            [[:as +compojure-api-request+] ?request__3574__auto__]
+                            (plumbing.core/letk
+                              [[field :- EXPENSIVE
+                                field2 {default :-, s/Int (inc 42)}
+                                :& foo :- {s/Keyword s/Keyword}
+                                :as all]
+                               (compojure.api.coercion/coerce-request!
+                                 ?form-params-schema108882
+                                 :form-params
+                                 :string
+                                 true
+                                 false
+                                 +compojure-api-request+)]
+                              (do (ok "kikka"))))))}))))
   (testing "no context"
     (let [times (atom {})
           record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
@@ -1305,100 +1309,126 @@
       (dorun (repeatedly 10 exercise))
       (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 11} @times))))
   (testing "inferred static context"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" []
                   (GET "/ping" []
-                       :body [body (do (swap! times inc) s/Any)]
+                       :form-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 1 @times))))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 11} @times))))
   (testing "dynamic context that doesn't bind variables"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" []
                   :dynamic true
                   (GET "/ping" []
-                       :body [body (do (swap! times inc) s/Any)]
+                       :form-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 11 @times))))
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "dynamic context where schema is bound outside context"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (let [s s/Any]
                   (context
                     "" []
                     :dynamic true
                     (GET "/ping" []
                          ;;TODO could lift this since the locals occur outside the context
-                         :body [body (do (swap! times inc) s)]
+                         :form-params [field :- (record :field s/Str)
+                                       field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                       & foo :- {(record :extra-keys s/Keyword)
+                                                 (record :extra-vals s/Keyword)} :as all]
                          (ok "kikka"))))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 11 @times))))
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "dynamic context that binds req and uses it in schema"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" req
                   (GET "/ping" req
-                       :body [body (do (swap! times inc)
-                                       ;; should never lift this since it refers to request
-                                       (second [req s/Any]))]
+                       :form-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 11 @times))))
-  (testing "bind :body in static context"
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
+  (testing "bind :form-params in static context"
     (is-thrown-with-msg?
       AssertionError
       #"cannot be :static"
       (eval `(context
                "" []
                :static true
-               :body [body (do (swap! times update :outer inc)
-                               s/Any)]
+               :form-params [field :- (record :field s/Str)
+                             field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                             & foo :- {(record :extra-keys s/Keyword)
+                                       (record :extra-vals s/Keyword)} :as all]
                (GET "/ping" req
                     :body [body (do (swap! times update :inner inc)
                                     s/Any)]
                     (ok "kikka"))))))
-  (testing "bind :body in dynamic context"
-    (let [times (atom {:outer 0 :inner 0})
+  (testing "bind :form-params in dynamic context"
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (context
                   "" []
                   :dynamic true
-                  :body [body (do (swap! times update :outer inc)
-                                  s/Any)]
                   (GET "/ping" req
-                       :body [body (do (swap! times update :inner inc)
-                                       s/Any)]
+                       :form-params [field :- (record :field s/Str)
+                                     field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                     & foo :- {(record :extra-keys s/Keyword)
+                                               (record :extra-vals s/Keyword)} :as all]
                        (ok "kikka")))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {} @times))
       (exercise)
-      (is (= {:outer 1 :inner 1} @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= {:outer 1 :inner 11} @times))))
+      (is (= {:field 11 :default 11 :extra-keys 11 :extra-vals 11  :default-never 11} @times))))
   (testing "idea for lifting impl"
-    (let [times (atom 0)
+    (let [times (atom {})
+          record (fn [path schema] (swap! times update path (fnil inc 0)) schema)
           route (let [rs (GET "/ping" req
-                              :body [body (do (swap! times inc)
-                                              s/Any)]
+                              :form-params [field :- (record :field s/Str)
+                                            field2 {default :- (record :default s/Int) (record :default-never (inc 42))}
+                                            & foo :- {(record :extra-keys s/Keyword)
+                                                      (record :extra-vals s/Keyword)} :as all]
                               (ok "kikka"))]
                   (context
                     "" []
                     :dynamic true
                     rs))
-          exercise #(is (= "kikka" (:body (route {:request-method :get :uri "/ping"}))))]
+          exercise #(is (= "kikka" (:body (route {:form-params {:field "a" :field2 2} :request-method :get :uri "/ping"}))))]
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1} @times))
       (exercise)
-      (is (= 1 @times))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 1} @times))
       (dorun (repeatedly 10 exercise))
-      (is (= 1 @times)))))
+      (is (= {:field 1 :default 1 :extra-keys 1 :extra-vals 1 :default-never 11} @times)))))